+++ src/so_vits_svc_fork/gui.py
@@ -25,6 +25,13 @@ def play_audio(path: Path | str):
 def main():
     sg.theme("Dark")
     model_candidates = list(sorted(Path("./logs/44k/").glob("G_*.pth")))
+
+    devices = sd.query_devices()
+    input_devices = [d["name"] for d in devices if d["max_input_channels"] > 0]
+    output_devices = [d["name"] for d in devices if d["max_output_channels"] > 0]
+    devices[sd.default.device[0]]["name"]
+    devices[sd.default.device[1]]["name"]
+
     layout = [
         [
             sg.Frame(
@@ -35,14 +42,16 @@ def main():
                         sg.Push(),
                         sg.InputText(
                             key="model_path",
-                            default_text=model_candidates[-1].as_posix()
+                            default_text=model_candidates[-1].absolute().as_posix()
                             if model_candidates
                             else "",
                         ),
                         sg.FileBrowse(
-                            initial_folder="./logs/44k/"
+                            initial_folder=Path("./logs/44k/").absolute
                             if Path("./logs/44k/").exists()
-                            else "."
+                            else Path(".").absolute().as_posix(),
+                            key="model_path_browse",
+                            file_types=(("PyTorch", "*.pth"),),
                         ),
                     ],
                     [
@@ -50,20 +59,32 @@ def main():
                         sg.Push(),
                         sg.InputText(
                             key="config_path",
-                            default_text="./configs/44k/config.json",
+                            default_text=Path("./configs/44k/config.json")
+                            .absolute()
+                            .as_posix()
+                            if Path("./configs/44k/config.json").exists()
+                            else "",
                             enable_events=True,
                         ),
                         sg.FileBrowse(
-                            initial_folder="./configs/44k/"
+                            initial_folder=Path("./configs/44k/").as_posix()
                             if Path("./configs/44k/").exists()
-                            else "."
+                            else Path(".").absolute().as_posix(),
+                            key="config_path_browse",
+                            file_types=(("JSON", "*.json"),),
                         ),
                     ],
                     [
                         sg.Text("Cluster model path"),
                         sg.Push(),
                         sg.InputText(key="cluster_model_path"),
-                        sg.FileBrowse(),
+                        sg.FileBrowse(
+                            initial_folder="./logs/44k/"
+                            if Path("./logs/44k/").exists()
+                            else ".",
+                            key="cluster_model_path_browse",
+                            file_types=(("PyTorch", "*.pth"),),
+                        ),
                     ],
                 ],
             )
@@ -207,6 +228,24 @@ def main():
                             key="realtime_algorithm",
                         ),
                     ],
+                    [
+                        sg.Text("Input device"),
+                        sg.Combo(
+                            key="input_device",
+                            values=input_devices,
+                            size=(20, 1),
+                            default_value=input_devices[0],
+                        ),
+                    ],
+                    [
+                        sg.Text("Output device"),
+                        sg.Combo(
+                            key="output_device",
+                            values=output_devices,
+                            size=(20, 1),
+                            default_value=output_devices[0],
+                        ),
+                    ],
                 ],
             )
         ],
@@ -234,7 +273,8 @@ def main():
             def update_combo() -> None:
                 from . import utils
 
-                if Path(values["config_path"]).exists():
+                config_path = Path(values["config_path"])
+                if config_path.exists() and config_path.is_file():
                     hp = utils.get_hparams_from_file(values["config_path"])
                     LOG.info(f"Loaded config from {values['config_path']}")
                     window["speaker"].update(
@@ -245,7 +285,16 @@ def main():
                 LOG.info(f"Event {event}, values {values}")
             if values["speaker"] == "":
                 update_combo()
-
+            if event.endswith("_path"):
+                browser = window[f"{event}_browse"]
+                if isinstance(browser, sg.Button):
+                    LOG.info(
+                        f"Updating browser {browser} to {Path(values[event]).parent}"
+                    )
+                    browser.InitialFolder = Path(values[event]).parent
+                    browser.update()
+                else:
+                    LOG.warning(f"Browser {browser} is not a FileBrowse")
             if event == "config_path":
                 update_combo()
             elif event == "infer":
@@ -308,6 +357,8 @@ def main():
                         version=int(values["realtime_algorithm"][0]),
                         device="cuda" if values["use_gpu"] else "cpu",
                         block_seconds=values["block_seconds"],
+                        input_device=values["input_device"],
+                        output_device=values["output_device"],
                     ),
                 )
             elif event == "stop_vc":